fix(NetworkClient): decode snake case
