--- ImageFeed/Extension UITableViewDelegate.swift
@@ -1,31 +1,0 @@
-//
-//  extension UITableViewDelegate.swift
-//  ImageFeed
-//
-//  Created by ulyana on 28.11.24.
-//
-
-import Foundation
-import UIKit
-
-extension ImagesListViewController: UITableViewDelegate {
-    
-    /// переход от таблице к одной картинке
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        performSegue(withIdentifier: showSingleImageSegueIdentifier, sender: indexPath)
-    }
-
-    /// просчет высоты ячейки
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        guard let image = UIImage(named: photosName[indexPath.row]) else {
-            return 0
-        }
-        
-        let imageInsets = UIEdgeInsets(top: 4, left: 16, bottom: 4, right: 16)
-        let imageViewWidth = tableView.bounds.width - imageInsets.left - imageInsets.right
-        let imageWidth = image.size.width
-        let scale = imageViewWidth / imageWidth
-        let cellHeight = image.size.height * scale + imageInsets.top + imageInsets.bottom
-        return cellHeight
-    }
-}
