+++ ImageFeed/OAuthTokenResponseBody.swift
@@ -9,8 +9,4 @@ import UIKit
 
 struct OAuthTokenResponseBody: Decodable {
     let accessToken: String
-    
-    private enum CodingKeys: String, CodingKey {
-        case accessToken = "access_token"
-    }
 }
