+++ ImageFeed/Profile/Models/ProfileResult.swift
@@ -12,10 +12,4 @@ struct ProfileResult: Decodable {
     let firstName: String
     let lastName: String?
     let bio: String?
-    
-    enum CodingKeys: String, CodingKey {
-        case username, bio
-        case firstName = "first_name"
-        case lastName = "last_name"
-    }
 }
