+++ ImageFeed/Profile/Models/UserResult.swift
@@ -9,8 +9,4 @@ import Foundation
 
 struct UserResult: Decodable {
     let profileImage: ProfileImage
-    
-    enum CodingKeys: String, CodingKey {
-        case profileImage = "profile_image"
-    }
 }
