--- ImageFeed/extension ImagesListViewController.swift
@@ -1,25 +1,0 @@
-//
-//  File.swift
-//  ImageFeed
-//
-//  Created by ulyana on 1.12.24.
-//
-
-import Foundation
-import UIKit
-
-extension ImagesListViewController {
-    
-    func configCell(for cell: ImagesListCell, with indexPath: IndexPath) {
-        guard let image = UIImage(named: photosName[indexPath.row]) else {
-            return
-        }
-
-        cell.cellImage.image = image
-        cell.dateLabel.text = dateFormatter.string(from: Date())
-
-        let isLiked = indexPath.row % 2 == 0
-        let likeImage = isLiked ? UIImage(named: "like_button_on") : UIImage(named: "like_button_off")
-        cell.likeButton.setImage(likeImage, for: .normal)
-    }
-}
